Fix Java 17 error.

+++ net.solarnetwork.node.datum.bacnet.test/src/net/solarnetwork/node/datum/bacnet/test/BacnetDatumDataSourceConfigCsvParserTests.java
@@ -1,21 +1,21 @@
 /* ==================================================================
  * BacnetDatumDataSourceConfigCsvParserTests.java - 9/11/2022 2:42:31 pm
- * 
+ *
  * Copyright 2022 SolarNetwork.net Dev Team
- * 
- * This program is free software; you can redistribute it and/or 
- * modify it under the terms of the GNU General Public License as 
- * published by the Free Software Foundation; either version 2 of 
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation; either version 2 of
  * the License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful, 
- * but WITHOUT ANY WARRANTY; without even the implied warranty of 
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License 
- * along with this program; if not, write to the Free Software 
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
  * 02111-1307 USA
  * ==================================================================
  */
@@ -53,7 +53,7 @@ import net.solarnetwork.util.ByteUtils;
 
 /**
  * Test cases for the {@link BacnetDatumDataSourceConfigCsvParser} class.
- * 
+ *
  * @author matt
  * @version 1.0
  */
@@ -230,11 +230,11 @@ public class BacnetDatumDataSourceConfigCsvParserTests {
 			Float covIncrement, String datumPropName, DatumSamplesType datumPropType, Integer mult,
 			int scale) {
 		assertThat(format("Prop config %s object type", msg), propConfig.getObjectType(),
-				is(equalTo(objectType != null ? objectType.getCode() : null)));
+				is(equalTo(objectType != null ? (Integer) objectType.getCode() : null)));
 		assertThat(format("Prop config %s object number", msg), propConfig.getObjectNumber(),
 				is(equalTo(objectNumber)));
 		assertThat(format("Prop config %s property type", msg), propConfig.getPropertyId(),
-				is(equalTo(propType != null ? propType.getCode() : null)));
+				is(equalTo(propType != null ? (Integer) propType.getCode() : null)));
 		assertThat(format("Prop config %s COV increment", msg), propConfig.getCovIncrement(),
 				is(equalTo(covIncrement)));
 		assertThat(format("Prop config %s datum property type", msg), propConfig.getPropertyType(),
